+++ src/common/functions.ts
@@ -1,5 +1,9 @@
 import React from 'react'
 
+export const capitalise = (word: string): string => {
+  return word ? word.substring(0,1).toUpperCase() + word.substring(1).toLowerCase() : ''
+}
+
 export const preventReloadByEnter = (event: React.KeyboardEvent): boolean => {
   const isAction = event.key === 'Enter'
   if (isAction) {